Finally fixed the problem with IntervalSearch
Didn't realize that the issue was, that CoeJudge only called the specific function, and therefore didn't run the check to see, if the interval was written the wrong way around. I moved the check to the function (where it should have been all along), and added a little logic to make calculations faster - though in hindsight, the switch-statement was not really needed, since I'd just use a break to move on after finding a valid value anyway, which wouldn't usually take too long.