Solved problem 3.a
I removed that round-about solution, trat utilized the maximum distance as if it were a distance between the two beasts.
I realized that the task was to pick values in the interval [-s;s] so that the beast moves "a maximal distance towards the prey". Since the distance itself doesn't have to be in the interval, the numbers are added to the coords in order for the distance moved to be as long as possible, towards the prey.

Found a problem in which the prey doesn't add the same value to each coordinate - ex: moves from [-2;10] to [1;11] meaning it added the values 3 and 1...
Strange...